refactor: change impoer quotes, move user to folder
