+++ src/err/index.ts
@@ -1,3 +1,3 @@
 
-export * from './not-found';
-export * from './validation-err';
+export * from "./not-found";
+export * from "./validation-error";
